Move contents of OutageConcern into IdvStepConcern
Since it was only two methods and that was the only place it was included.

[skip changelog]
